Nvs/report service

See merge request tantd1/microservices-core-banking!88

+++ banking-report-service/src/main/java/com/ctv_it/klb/dto/baseInfo/AccountInfoDTO.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 @ToString
 public class AccountInfoDTO {
 
-  private int no;
+  private long no;
   private Long id;
   private String type;
   private String name;
